Refactor: weekItem ui 수정
