+++ FrontEnd/DDIB/.env
@@ -8,3 +8,5 @@ KAKAO_CLIENT_SECRET="oqFU150X8k60LpSswOxuTanX6cCZsBWr"
 
 NEXTAUTH_SECRET="aBZ5Js2YS1pA6/ayJ8bYQelaaFUYnFln+mkGvvazSJ4="
 NEXT_PUBLIC_USE_MSW = true
+
+NEXT_PUBLIC_MAINTENANCE = false
