+++ FrontEnd/DDIB/package.json
@@ -7,7 +7,7 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "mock": "npx tsx watch ./src/_mocks/http.ts"
+    "mock": "npx tsx watch ./src/mocks/http.ts"
   },
   "dependencies": {
     "@faker-js/faker": "^8.4.1",
