+++ FrontEnd/DDIB/src/mocks/browser.ts
