+++ FrontEnd/DDIB/src/mocks/productHandlers.ts
